requirements.txt
streamlit
openai
ffmpeg-python
python-dotenv
requests
Pillow
loguru 